LayoutTests/fast/forms/password-doubleclick-selection.html

<html>
<head>
    <script>
    async function test() {
        var pass = document.getElementById('tf');
        var res = document.getElementById('res');

        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.waitUntilDone();
            await eventSender.asyncMouseMoveTo(45, 25);
            await eventSender.asyncMouseDown();
            await eventSender.asyncMouseUp();
            await eventSender.asyncMouseDown();
            await eventSender.asyncMouseUp();
            if (pass.selectionStart == 0 && pass.selectionEnd == 11)
                res.innerHTML = "Test Passed.";
            else
                res.innerHTML = "Test Failed.  SelectionStart index = " + pass.selectionStart + ".  SelectionEnd index = " + pass.selectionEnd + ".";
            testRunner.notifyDone();
        }
    }
    </script>
</head>
<body onload="test()">
    <input type="password" value="123 456 789" id="tf"></input>
    This tests that double clicking in a password field will not expose word boundaries.<br>
    <br><div id="res"></div>
</body>
</html>
